(BSR)[PRO] refactor: move maintainance reducer to TS
