--- pro/src/store/selectors/data/__specs__/mockState.json
@@ -1,207 +1,0 @@
-{
-  "card": {
-    "areDetailsVisible": false,
-    "draggable": true,
-    "isActive": false,
-    "unFlippable": false
-  },
-  "data": {
-    "bookings": [],
-    "offerers": [
-        {
-          "id": "4Q",
-          "address": "RUE DES SAPOTILLES",
-          "city": "Cayenne",
-          "dateCreated": "2019-05-06T09:12:46.278743Z",
-          "dateModifiedAtLastProvider": "2019-05-06T09:13:08.458343Z",
-          "isActive": true,
-          "isValidated": true,
-          "lastProviderId": null,
-          "modelName": "Offerer",
-          "nOffers": 5,
-          "name": "Bar des amis",
-          "postalCode": "97300",
-          "siren": "222222233"
-      }],
-    "offers": [
-      {
-        "id": "AXBQ",
-        "bookingEmail": null,
-        "dateCreated": "2018-11-13T09:54:56.103080Z",
-        "dateModifiedAtLastProvider": "2018-11-13T09:54:56.103073Z",
-        "isEvent": false,
-        "isActive": true,
-        "lastProviderId": null,
-        "modelName": "Offer",
-        "productId": "V4",
-        "isThing": true,
-        "venueId": "AYJA",
-        "eventOccurrencesIds": [],
-        "mediationsIds": [],
-        "stocksIds": ["DCRA"]
-      }
-    ],
-    "readRecommendations": [
-      {
-        "id": "AY9A",
-        "dateRead": "2019-05-06T09:13:45.266Z"
-      },
-      {
-        "id": "AY9Q",
-        "dateRead": "2019-05-06T09:13:46.910Z"
-      },
-      {
-        "id": "AYKQ",
-        "dateRead": "2019-05-06T09:20:59.025Z"
-      }
-    ],
-    "recommendations": [
-      {
-        "id": "A2KA",
-        "dateCreated": "2019-05-06T09:40:11.809052Z",
-        "dateRead": null,
-        "dateUpdated": "2019-05-06T09:40:11.809076Z",
-        "isClicked": true,
-        "isFavorite": false,
-        "isFirst": false,
-        "mediation": {
-          "authorId": "AK2A",
-          "backText": null,
-          "credit": "undefined",
-          "dateCreated": "2019-05-06T09:40:00.408589Z",
-          "dateModifiedAtLastProvider": "2019-05-06T09:40:00.408532Z",
-          "frontText": null,
-          "id": "AMCQ",
-          "isActive": true,
-          "lastProviderId": null,
-          "modelName": "Mediation",
-          "offerId": "AX8Q",
-          "tutoIndex": null
-        },
-        "mediationId": "AMCQ",
-        "modelName": "Recommendation",
-        "offer": {
-          "ageMax": null,
-          "ageMin": null,
-          "bookingEmail": "fake@email.com",
-          "conditions": null,
-          "dateCreated": "2019-05-06T09:39:21.385332Z",
-          "dateModifiedAtLastProvider": "2019-05-06T09:39:21.385288Z",
-          "description": "Raoul Taburin, c\"est l\"histoire d\"un petit garçon devenu grand sans savoir faire du vélo. L\"histoire d\"un immense malentendu vécu comme une malédiction. Un imposteur malgré lui.",
-          "durationMinutes": 120,
-          "extraData": {
-            "author": "Michal Aviad",
-            "stageDirector": "Laurne Valik",
-            "visa": "5454VCJGFDJRT"
-          },
-          "id": "AX8Q",
-          "isActive": true,
-          "isEvent": true,
-          "isFinished": false,
-          "isFullyBooked": false,
-          "isNational": true,
-          "isThing": false,
-          "lastProviderId": null,
-          "mediaUrls": [],
-          "modelName": "Offer",
-          "name": "Nice offre",
-          "product": {
-            "ageMax": null,
-            "ageMin": null,
-            "conditions": null,
-            "dateModifiedAtLastProvider": "2019-05-06T09:39:21.350837Z",
-            "description": "Raoul Taburin, c\"est l\"histoire d\"un petit garçon devenu grand sans savoir faire du vélo. L\"histoire d\"un immense malentendu vécu comme une malédiction. Un imposteur malgré lui.",
-            "durationMinutes": 120,
-            "extraData": {
-              "author": "Michal Aviad",
-              "stageDirector": "Laurne Valik",
-              "visa": "5454VCJGFDJRT"
-            },
-            "id": "BAUQ",
-            "isNational": true,
-            "lastProviderId": null,
-            "mediaUrls": [],
-            "modelName": "Product",
-            "name": "Nice offre",
-            "owningOffererId": "4Y",
-            "thumbUrl": "http://localhost/storage/thumbs/products/BAUQ",
-            "url": null
-          },
-          "productId": "BAUQ",
-          "stocks": [
-            {
-              "available": null,
-              "beginningDatetime": "2019-05-26T09:39:24Z",
-              "bookingLimitDatetime": "2019-05-17T09:39:24Z",
-              "bookingRecapSent": null,
-              "dateModified": "2019-05-06T09:39:34.241354Z",
-              "dateModifiedAtLastProvider": "2019-05-06T09:39:34.241239Z",
-              "id": "AWNQ",
-              "isSoftDeleted": false,
-              "lastProviderId": null,
-              "modelName": "Stock",
-              "offerId": "AX8Q",
-              "price": 0
-            },
-            {
-              "available": null,
-              "beginningDatetime": "2019-05-27T09:39:24Z",
-              "bookingLimitDatetime": "2019-05-25T09:39:24Z",
-              "bookingRecapSent": null,
-              "dateModified": "2019-05-06T09:39:40.773023Z",
-              "dateModifiedAtLastProvider": "2019-05-06T09:39:40.772850Z",
-              "id": "AW8A",
-              "isSoftDeleted": false,
-              "lastProviderId": null,
-              "modelName": "Stock",
-              "offerId": "AX8Q",
-              "price": 0
-            }
-          ],
-          "thingId": null,
-          "url": null,
-          "venue": {
-            "address": "RUE NIEUPORT",
-            "bookingEmail": "fake@email.com",
-            "city": "Drancy",
-            "comment": null,
-            "dateModifiedAtLastProvider": "2019-05-06T09:13:13.134271Z",
-            "departementCode": "93",
-            "id": "AFCA",
-            "isVirtual": false,
-            "lastProviderId": null,
-            "latitude": 48.91678,
-            "longitude": 2.43846,
-            "managingOfferer": {
-              "address": "RUE NIEUPORT",
-              "city": "Drancy",
-              "dateCreated": "2019-05-06T09:12:46.278743Z",
-              "dateModifiedAtLastProvider": "2019-05-06T09:13:08.463068Z",
-              "id": "4Y",
-              "isActive": true,
-              "lastProviderId": null,
-              "modelName": "Offerer",
-              "name": "Club Dorothy",
-              "postalCode": "93700",
-              "siren": "222222223",
-              "validationToken": null
-            },
-            "managingOffererId": "4Y",
-            "modelName": "Venue",
-            "name": "Maison de la Brique",
-            "postalCode": "93700",
-            "publicName": null,
-            "siret": "22222222311111",
-            "validationToken": null
-          },
-          "venueId": "AFCA"
-        },
-        "offerId": "AX8Q",
-        "search": "keywords_string=nice",
-        "shareMedium": null,
-        "thumbUrl": "http://localhost/storage/thumbs/mediations/AMCQ",
-        "userId": "AK4A"
-      }
-    ]
-  }
-}
