--- pro/src/store/selectors/maintenanceSelector.js
@@ -1,5 +1,0 @@
-export function maintenanceSelector(state) {
-  return state.maintenance.isActivated
-}
-
-export default maintenanceSelector
