+++ pro/src/store/selectors/maintenanceSelector.ts
@@ -1,0 +1,5 @@
+import { RootState } from 'store/reducers'
+
+export function maintenanceSelector(state: RootState) {
+  return state.maintenance.isActivated
+}
